+++ NEWS.md
@@ -1,5 +1,5 @@
 # reutils 0.1.0.99
 
-* new return type `content(x, as = "textConnection")`.
+* Add global option `reutils.show.headlines` to control the number of head lines displayed in show,efetch method (currently only if `rettype = "text"`.
 
-* move to testthat 0.8
+* new return type `content(x, as = "textConnection")`.